feat: move module details to cursor-following hover tooltip
Extract Module Details from the right InfoPanel into a floating tooltip
that follows the mouse cursor over the canvas. The tooltip appears
immediately with no delay, grows freely to fit content, and repositions
to stay within viewport bounds.

Also increases Content section max-height (max-h-32 → max-h-64) and
character truncation limit (300 → 600).

Closes #1
